начал делать возможность привязку нового объекта к уже существующим

+++ src/main/java/itmo/app/controller/MovieController.java
@@ -1,7 +1,12 @@
 package itmo.app.controller;
 
+import itmo.app.model.entity.Coordinates;
 import itmo.app.model.entity.Movie;
+import itmo.app.model.entity.Person;
+import itmo.app.model.repository.CoordinatesRepository;
 import itmo.app.model.repository.MovieRepository;
+import itmo.app.model.repository.PersonRepository;
+import org.hibernate.PersistentObjectException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,8 +22,18 @@ public class MovieController {
 	@Autowired
 	private MovieRepository movieRepository;
 	
+	@Autowired
+	private PersonRepository personRepository;
+	
+	@Autowired
+	private CoordinatesRepository coordinatesRepository;
+	
 	@PostMapping("/action")
 	public ResponseEntity<Movie> createMovie(@RequestBody Movie movie) {
+		Optional<Coordinates> existingCoordinates = coordinatesRepository.findById(movie.getCoordinates().getId());
+		
+		existingCoordinates.ifPresent(movie::setCoordinates);
+		
 		Movie newMovie = movieRepository.save(movie);
 		return new ResponseEntity<>(newMovie, HttpStatus.CREATED);
 	}
@@ -30,9 +45,21 @@ public class MovieController {
 	
 	@DeleteMapping("/action/{id}")
 	public ResponseEntity<List<Movie>> deleteMovie(@PathVariable long id) {
-		movieRepository.deleteById(id);
+		Optional<Movie> movieToDelete = movieRepository.findById(id);
+		
+		if (movieToDelete.isPresent()) {
+			Coordinates coordinates = movieToDelete.get().getCoordinates();
+			movieRepository.deleteById(id);
+			
+			List<Movie> moviesWithSameCoordinates = movieRepository.findByCoordinates(coordinates);
+			if (moviesWithSameCoordinates.isEmpty()) {
+				coordinatesRepository.delete(coordinates);
+			}
+		}
+		
 		return new ResponseEntity<>(movieRepository.findAll(), HttpStatus.OK);
 	}
+
 	
 	@PutMapping("/action")
 	public ResponseEntity<List<Movie>> updateMovie(@RequestBody Movie movie) {
@@ -44,5 +71,18 @@ public class MovieController {
 	public ResponseEntity<Movie> getMovieById(@PathVariable long id) {
 		Optional<Movie> movie = movieRepository.findById(id);
 		return movie.map(value -> new ResponseEntity<>(value, HttpStatus.OK)).orElse(null);
+	}
+	
+	@GetMapping("/coordinates")
+	public ResponseEntity<List<Coordinates>> getAllCoordinates() {
+		// Здесь будет запрос к репозиторию координат (если он есть)
+		List<Coordinates> coordinates = coordinatesRepository.findAll();
+		return new ResponseEntity<>(coordinates, HttpStatus.OK);
+	}
+	
+	@GetMapping("/persons")
+	public ResponseEntity<List<Person>> getAllPersons() {
+		List<Person> persons = personRepository.findAll();
+		return new ResponseEntity<>(persons, HttpStatus.OK);
 	}
 }
